Memory Mapping Gates (unfinished)
QuantumCircuit holds the master state for the system, using a memory
pool to hold the gates (which are all reserved with proper sizing).
Parser still needs to be updated to reflect new changes. Additionally,
I'm going to use an AST to do parsing properly. This new design is far
more scalable while maintaining memory benefits.