Lookup Table + Operation Pool for Gates
Idea is to have a lookup table for qiskit circuit names and mapping them
onto Handles within the pool. There's an issue with the current
implementation which is detailed in the TODO for the new() fn. This is
fine for a temporary solution, but there should be a better way to
implement it